Fix to the wrong MPI_Reduce type in diags

+++ src/mpi_interface.c
@@ -299,33 +299,33 @@ void mpi_gather_diag(diag_data* data, int ntotal, int mpi_rank, int mpi_size,
         MPI_Reduce(
             mpi_rank == mpi_root ? MPI_IN_PLACE : data->dist5D.histogram,
             data->dist5D.histogram, data->dist5D.step_6 * data->dist5D.n_r,
-            MPI_INT, MPI_SUM, 0, MPI_COMM_WORLD);
+            MPI_DOUBLE, MPI_SUM, 0, MPI_COMM_WORLD);
     }
     if(data->dist6D_collect) {
         MPI_Reduce(
             mpi_rank == mpi_root ? MPI_IN_PLACE : data->dist6D.histogram,
             data->dist6D.histogram, data->dist6D.step_7 * data->dist6D.n_r,
-            MPI_INT, MPI_SUM, 0, MPI_COMM_WORLD);
+            MPI_DOUBLE, MPI_SUM, 0, MPI_COMM_WORLD);
     }
     if(data->distrho5D_collect) {
         MPI_Reduce(
             mpi_rank == mpi_root ? MPI_IN_PLACE : data->distrho5D.histogram,
             data->distrho5D.histogram,
             data->distrho5D.step_6 * data->distrho5D.n_rho,
-            MPI_INT, MPI_SUM, 0, MPI_COMM_WORLD);
+            MPI_DOUBLE, MPI_SUM, 0, MPI_COMM_WORLD);
     }
     if(data->distrho6D_collect) {
         MPI_Reduce(
             mpi_rank == mpi_root ? MPI_IN_PLACE : data->distrho6D.histogram,
             data->distrho6D.histogram,
             data->distrho6D.step_7*data->distrho6D.n_rho,
-            MPI_INT, MPI_SUM, 0, MPI_COMM_WORLD);
+            MPI_DOUBLE, MPI_SUM, 0, MPI_COMM_WORLD);
     }
     if(data->distCOM_collect) {
         MPI_Reduce(
             mpi_rank == mpi_root ? MPI_IN_PLACE : data->distCOM.histogram,
             data->distCOM.histogram, data->distCOM.step_2 * data->distCOM.n_mu,
-            MPI_INT, MPI_SUM, 0, MPI_COMM_WORLD);
+            MPI_DOUBLE, MPI_SUM, 0, MPI_COMM_WORLD);
     }
 
     if(data->diagorb_collect) {
